Add aliases for primary theme folders
