+++ styles/components/toggleButton.ts
@@ -1,0 +1,28 @@
+import { Theme } from "@/theme"
+import {
+    ContainedIcon,
+    InteractiveContainer,
+    InteractiveToggleableContainer,
+} from "@/theme/container"
+import { iconButton } from "@components/button"
+import { buttonWithIconStyle } from "."
+
+// Use ContainedIcon just as an example
+export function toggleButton(
+    theme: Theme
+): InteractiveToggleableContainer<InteractiveContainer<ContainedIcon>> {
+    const inactive = iconButton(theme)
+    const active = buttonWithIconStyle({
+        theme,
+        inputIntensity: {
+            bg: 32,
+            border: [48, 36],
+            fg: 100,
+        },
+    })
+
+    return {
+        inactive,
+        active,
+    }
+}
